Evaluate correlation of DNA methylation region and target gene expression

This function evaluate the correlation of the DNA methylation region and target gene expression using spearman rank correlation test. Note that genes with RNA expression equal to 0 for all samples will not be evaluated.

cor_region_dnam_target_gene(
  links,
  dnam,
  exp,
  filter.results = TRUE,
  min.cor.pval = 0.05,
  min.cor.estimate = 0,
  file.out,
  cores = 1
)

Arguments

links
A dataframe with the following columns: regionID (DNA methylation) and target (target gene)

dnam
DNA methylation matrix with regions in rows and samples in columns are samples. Samples should be in the same order as gene expression matrix (exp).

exp
Gene expression matrix (rows are genes, columns are samples) log2-normalized (log2(exp + 1)). Samples should be in the same order as the DNA methylation matrix.

filter.results
Filter results using min.cor.pval and min.cor.estimate thresholds

min.cor.pval
P-value threshold filter (default: 0.05)

min.cor.estimate
Correlation estimate threshold filter (default: not applied)

file.out
A csv file name which will be used to save the results.

cores
Number of CPU cores to be used. Default 1.

Value

A data frame with the following information: regionID, target gene, correlation pvalue and estimate between DNA methylation and target gene expression, FDR corrected p-values.

Examples

dnam <- t(matrix(sort(c(runif(20))), ncol = 1))
rownames(dnam) <- c("chr3:203727581-203728580")
colnames(dnam) <- paste0("Samples",1:20)
exp <- dnam
rownames(exp) <- c("ENSG00000232886")
colnames(exp) <- paste0("Samples",1:20)

links <- data.frame(
   "regionID" =  c("chr3:203727581-203728580"),
   "target" = "ENSG00000232886"
)

# Correlated DNAm and gene expression, display only significant associations
results.cor.pos <- cor_region_dnam_target_gene(
   links = links,
   dnam = dnam,
   exp = exp,
   filter.results = TRUE,
   min.cor.pval = 0.05,
   min.cor.estimate = 0.0
)

#> Removing genes with RNA expression equal to 0/NA for all samples

#> Removing 0 out of 1 genes

#> Removing regions with beta-values equal to NA for all samples

#> |                                                    |  0%                      |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s 

if (FALSE) {
# Load data
data("gene.exp.chr21.log2")
data("dna.met.chr21")
dna.met.chr21 <- make_se_from_dnam_probes(dna.met.chr21)

# Map example region to closest gene
links <- get_region_target_gene(
  regions.gr = dna.met.chr21,
  genome = "hg19",
  method = "closest.gene"
)

# Samples in met and exp datasets should be in the same order.
identical (colnames (dna.met.chr21), colnames(gene.exp.chr21.log2))

# Correalted DNAm and gene expression, display only significant associations
results <- cor_region_dnam_target_gene(
  links = links,
  dnam = dna.met.chr21,
  exp = gene.exp.chr21.log2
)

# display all associations
results.all <- cor_region_dnam_target_gene(
   links = links,
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2,
   filter.results = FALSE
)
}