Summarize promoter methylation beta values by mean or median.

First, identify gene promoter regions (+-2Kkb around TSS). When promoters overlap they will be merged. Then, for each promoter region calculate the mean DNA methylation of probes overlapping the region.

get_promoter_avg(
  dnam,
  genome,
  arrayType,
  cores = 1,
  upstream.dist.tss = 2000,
  downstream.dist.tss = 2000
)

Arguments

dnam
A DNA methylation matrix

genome
Human genome of reference hg38 or hg19

arrayType
DNA methylation array type (450k or EPIC)

cores
A integer number to use multiple cores. Default 1 core.

upstream.dist.tss
Number of base pairs (bp) upstream of TSS to consider as promoter regions

downstream.dist.tss
Number of base pairs (bp) downstream of TSS to consider as promoter regions

Value

A RangedSummarizedExperiment with promoter region and mean beta-values of CpGs within it.

Examples

data("dna.met.chr21")
promoter.avg <- get_promoter_avg(dna.met.chr21, genome = "hg19", arrayType = "450k")

#> o Get promoter regions for hg19

#> oo Number of promoter regions in hg19: 57642

#> o Get DNA methylation regions overlapping promoter regions

#> o Get mean DNA methylation of regions overlapping each promoter region

#> |                                                    |  0%                      |                                                    |0.3773585% ~1 s remaining |                                                    |0.754717% ~1 s remaining  |                                                    |1.132075% ~1 s remaining  |                                                    |1.509434% ~1 s remaining  |                                                    |1.886792% ~1 s remaining  |=                                                   |2.264151% ~1 s remaining  |=                                                   |2.641509% ~1 s remaining  |=                                                   |3.018868% ~1 s remaining  |=                                                   |3.396226% ~1 s remaining  |=                                                   |3.773585% ~1 s remaining  |==                                                  |4.150943% ~1 s remaining  |==                                                  |4.528302% ~1 s remaining  |==                                                  |4.90566% ~1 s remaining   |==                                                  |5.283019% ~1 s remaining  |==                                                  |5.660377% ~1 s remaining  |===                                                 |6.037736% ~1 s remaining  |===                                                 |6.415094% ~9 s remaining  |===                                                 |6.792453% ~8 s remaining  |===                                                 |7.169811% ~8 s remaining  |===                                                 |7.54717% ~7 s remaining   |====                                                |7.924528% ~7 s remaining  |====                                                |8.301887% ~7 s remaining  |====                                                |8.679245% ~6 s remaining  |====                                                |9.056604% ~6 s remaining  |====                                                |9.433962% ~6 s remaining  |=====                                               |9.811321% ~6 s remaining  |=====                                               |10.18868% ~6 s remaining  |=====                                               |10.56604% ~5 s remaining  |=====                                               |10.9434% ~5 s remaining   |=====                                               |11.32075% ~5 s remaining  |======                                              |11.69811% ~5 s remaining  |======                                              |12.07547% ~5 s remaining  |======                                              |12.45283% ~5 s remaining  |======                                              |12.83019% ~4 s remaining  |======                                              |13.20755% ~4 s remaining  |=======                                             |13.58491% ~4 s remaining  |=======                                             |13.96226% ~4 s remaining  |=======                                             |14.33962% ~4 s remaining  |=======                                             |14.71698% ~4 s remaining  |=======                                             |15.09434% ~4 s remaining  |========                                            |15.4717% ~4 s remaining   |========                                            |15.84906% ~4 s remaining  |========                                            |16.22642% ~4 s remaining  |========                                            |16.60377% ~3 s remaining  |========                                            |16.98113% ~3 s remaining  |=========                                           |17.35849% ~3 s remaining  |=========                                           |17.73585% ~3 s remaining  |=========                                           |18.11321% ~3 s remaining  |=========                                           |18.49057% ~3 s remaining  |=========                                           |18.86792% ~3 s remaining  |==========                                          |19.24528% ~3 s remaining  |==========                                          |19.62264% ~3 s remaining  |==========                                          | 20% ~3 s remaining       |==========                                          |20.37736% ~3 s remaining  |==========                                          |20.75472% ~3 s remaining  |==========                                          |21.13208% ~3 s remaining  |===========                                         |21.50943% ~3 s remaining  |===========                                         |21.88679% ~3 s remaining  |===========                                         |22.26415% ~3 s remaining  |===========                                         |22.64151% ~3 s remaining  |===========                                         |23.01887% ~2 s remaining  |============                                        |23.39623% ~2 s remaining  |============                                        |23.77358% ~2 s remaining  |============                                        |24.15094% ~2 s remaining  |============                                        |24.5283% ~2 s remaining   |============                                        |24.90566% ~2 s remaining  |=============                                       |25.28302% ~2 s remaining  |=============                                       |25.66038% ~2 s remaining  |=============                                       |26.03774% ~2 s remaining  |=============                                       |26.41509% ~2 s remaining  |=============                                       |26.79245% ~2 s remaining  |==============                                      |27.16981% ~2 s remaining  |==============                                      |27.54717% ~2 s remaining  |==============                                      |27.92453% ~2 s remaining  |==============                                      |28.30189% ~2 s remaining  |==============                                      |28.67925% ~2 s remaining  |===============                                     |29.0566% ~2 s remaining   |===============                                     |29.43396% ~2 s remaining  |===============                                     |29.81132% ~2 s remaining  |===============                                     |30.18868% ~2 s remaining  |===============                                     |30.56604% ~2 s remaining  |================                                    |30.9434% ~2 s remaining   |================                                    |31.32075% ~2 s remaining  |================                                    |31.69811% ~2 s remaining  |================                                    |32.07547% ~2 s remaining  |================                                    |32.45283% ~2 s remaining  |=================                                   |32.83019% ~2 s remaining  |=================                                   |33.20755% ~2 s remaining  |=================                                   |33.58491% ~2 s remaining  |=================                                   |33.96226% ~2 s remaining  |=================                                   |34.33962% ~2 s remaining  |==================                                  |34.71698% ~2 s remaining  |==================                                  |35.09434% ~2 s remaining  |==================                                  |35.4717% ~2 s remaining   |==================                                  |35.84906% ~1 s remaining  |==================                                  |36.22642% ~1 s remaining  |===================                                 |36.60377% ~1 s remaining  |===================                                 |36.98113% ~1 s remaining  |===================                                 |37.35849% ~1 s remaining  |===================                                 |37.73585% ~1 s remaining  |===================                                 |38.11321% ~1 s remaining  |====================                                |38.49057% ~1 s remaining  |====================                                |38.86792% ~1 s remaining  |====================                                |39.24528% ~1 s remaining  |====================                                |39.62264% ~1 s remaining  |====================                                | 40% ~1 s remaining       |====================                                |40.37736% ~1 s remaining  |=====================                               |40.75472% ~1 s remaining  |=====================                               |41.13208% ~1 s remaining  |=====================                               |41.50943% ~1 s remaining  |=====================                               |41.88679% ~1 s remaining  |=====================                               |42.26415% ~1 s remaining  |======================                              |42.64151% ~1 s remaining  |======================                              |43.01887% ~1 s remaining  |======================                              |43.39623% ~1 s remaining  |======================                              |43.77358% ~1 s remaining  |======================                              |44.15094% ~1 s remaining  |=======================                             |44.5283% ~1 s remaining   |=======================                             |44.90566% ~1 s remaining  |=======================                             |45.28302% ~1 s remaining  |=======================                             |45.66038% ~1 s remaining  |=======================                             |46.03774% ~1 s remaining  |========================                            |46.41509% ~1 s remaining  |========================                            |46.79245% ~1 s remaining  |========================                            |47.16981% ~1 s remaining  |========================                            |47.54717% ~1 s remaining  |========================                            |47.92453% ~1 s remaining  |=========================                           |48.30189% ~1 s remaining  |=========================                           |48.67925% ~1 s remaining  |=========================                           |49.0566% ~1 s remaining   |=========================                           |49.43396% ~1 s remaining  |=========================                           |49.81132% ~1 s remaining  |==========================                          |50.18868% ~1 s remaining  |==========================                          |50.56604% ~1 s remaining  |==========================                          |50.9434% ~1 s remaining   |==========================                          |51.32075% ~1 s remaining  |==========================                          |51.69811% ~1 s remaining  |===========================                         |52.07547% ~1 s remaining  |===========================                         |52.45283% ~1 s remaining  |===========================                         |52.83019% ~1 s remaining  |===========================                         |53.20755% ~1 s remaining  |===========================                         |53.58491% ~1 s remaining  |============================                        |53.96226% ~1 s remaining  |============================                        |54.33962% ~1 s remaining  |============================                        |54.71698% ~1 s remaining  |============================                        |55.09434% ~1 s remaining  |============================                        |55.4717% ~1 s remaining   |=============================                       |55.84906% ~1 s remaining  |=============================                       |56.22642% ~1 s remaining  |=============================                       |56.60377% ~1 s remaining  |=============================                       |56.98113% ~1 s remaining  |=============================                       |57.35849% ~1 s remaining  |==============================                      |57.73585% ~1 s remaining  |==============================                      |58.11321% ~1 s remaining  |==============================                      |58.49057% ~1 s remaining  |==============================                      |58.86792% ~1 s remaining  |==============================                      |59.24528% ~1 s remaining  |===============================                     |59.62264% ~1 s remaining  |===============================                     | 60% ~1 s remaining       |===============================                     |60.37736% ~1 s remaining  |===============================                     |60.75472% ~1 s remaining  |===============================                     |61.13208% ~1 s remaining  |===============================                     |61.50943% ~1 s remaining  |================================                    |61.88679% ~1 s remaining  |================================                    |62.26415% ~1 s remaining  |================================                    |62.64151% ~1 s remaining  |================================                    |63.01887% ~1 s remaining  |================================                    |63.39623% ~1 s remaining  |=================================                   |63.77358% ~1 s remaining  |=================================                   |64.15094% ~1 s remaining  |=================================                   |64.5283% ~1 s remaining   |=================================                   |64.90566% ~1 s remaining  |=================================                   |65.28302% ~1 s remaining  |==================================                  |65.66038% ~1 s remaining  |==================================                  |66.03774% ~1 s remaining  |==================================                  |66.41509% ~1 s remaining  |==================================                  |66.79245% ~1 s remaining  |==================================                  |67.16981% ~1 s remaining  |===================================                 |67.54717% ~1 s remaining  |===================================                 |67.92453% ~1 s remaining  |===================================                 |68.30189% ~1 s remaining  |===================================                 |68.67925% ~1 s remaining  |===================================                 |69.0566% ~1 s remaining   |====================================                |69.43396% ~1 s remaining  |====================================                |69.81132% ~1 s remaining  |====================================                |70.18868% ~1 s remaining  |====================================                |70.56604% ~1 s remaining  |====================================                |70.9434% ~1 s remaining   |=====================================               |71.32075% ~1 s remaining  |=====================================               |71.69811% ~0 s remaining  |=====================================               |72.07547% ~0 s remaining  |=====================================               |72.45283% ~0 s remaining  |=====================================               |72.83019% ~0 s remaining  |======================================              |73.20755% ~0 s remaining  |======================================              |73.58491% ~0 s remaining  |======================================              |73.96226% ~0 s remaining  |======================================              |74.33962% ~0 s remaining  |======================================              |74.71698% ~0 s remaining  |=======================================             |75.09434% ~0 s remaining  |=======================================             |75.4717% ~0 s remaining   |=======================================             |75.84906% ~0 s remaining  |=======================================             |76.22642% ~0 s remaining  |=======================================             |76.60377% ~0 s remaining  |========================================            |76.98113% ~0 s remaining  |========================================            |77.35849% ~0 s remaining  |========================================            |77.73585% ~0 s remaining  |========================================            |78.11321% ~0 s remaining  |========================================            |78.49057% ~0 s remaining  |=========================================           |78.86792% ~0 s remaining  |=========================================           |79.24528% ~0 s remaining  |=========================================           |79.62264% ~0 s remaining  |=========================================           | 80% ~0 s remaining       |=========================================           |80.37736% ~0 s remaining  |=========================================           |80.75472% ~0 s remaining  |==========================================          |81.13208% ~0 s remaining  |==========================================          |81.50943% ~0 s remaining  |==========================================          |81.88679% ~0 s remaining  |==========================================          |82.26415% ~0 s remaining  |==========================================          |82.64151% ~0 s remaining  |===========================================         |83.01887% ~0 s remaining  |===========================================         |83.39623% ~0 s remaining  |===========================================         |83.77358% ~0 s remaining  |===========================================         |84.15094% ~0 s remaining  |===========================================         |84.5283% ~0 s remaining   |============================================        |84.90566% ~0 s remaining  |============================================        |85.28302% ~0 s remaining  |============================================        |85.66038% ~0 s remaining  |============================================        |86.03774% ~0 s remaining  |============================================        |86.41509% ~0 s remaining  |=============================================       |86.79245% ~0 s remaining  |=============================================       |87.16981% ~0 s remaining  |=============================================       |87.54717% ~0 s remaining  |=============================================       |87.92453% ~0 s remaining  |=============================================       |88.30189% ~0 s remaining  |==============================================      |88.67925% ~0 s remaining  |==============================================      |89.0566% ~0 s remaining   |==============================================      |89.43396% ~0 s remaining  |==============================================      |89.81132% ~0 s remaining  |==============================================      |90.18868% ~0 s remaining  |===============================================     |90.56604% ~0 s remaining  |===============================================     |90.9434% ~0 s remaining   |===============================================     |91.32075% ~0 s remaining  |===============================================     |91.69811% ~0 s remaining  |===============================================     |92.07547% ~0 s remaining  |================================================    |92.45283% ~0 s remaining  |================================================    |92.83019% ~0 s remaining  |================================================    |93.20755% ~0 s remaining  |================================================    |93.58491% ~0 s remaining  |================================================    |93.96226% ~0 s remaining  |=================================================   |94.33962% ~0 s remaining  |=================================================   |94.71698% ~0 s remaining  |=================================================   |95.09434% ~0 s remaining  |=================================================   |95.4717% ~0 s remaining   |=================================================   |95.84906% ~0 s remaining  |==================================================  |96.22642% ~0 s remaining  |==================================================  |96.60377% ~0 s remaining  |==================================================  |96.98113% ~0 s remaining  |==================================================  |97.35849% ~0 s remaining  |==================================================  |97.73585% ~0 s remaining  |=================================================== |98.11321% ~0 s remaining  |=================================================== |98.49057% ~0 s remaining  |=================================================== |98.86792% ~0 s remaining  |=================================================== |99.24528% ~0 s remaining  |=================================================== |99.62264% ~0 s remaining  |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 1 s 

#> o Creating a SummarizedExperiment from DNA methylation input

#> oo Preparing SummarizedExperiment object