Obtain target genes of input regions

To map an input region to genes there are three options: 1) map region to closest gene 2) map region to all genes within a window around the region (default window.size = 500kbp (i.e. +/- 250kbp from start or end of the region)). 3) map region to a fixed number of nearby genes (upstream/downstream)

get_region_target_gene(
  regions.gr,
  genome = c("hg38", "hg19"),
  method = c("genes.promoter.overlap", "window", "nearby.genes", "closest.gene"),
  promoter.upstream.dist.tss = 2000,
  promoter.downstream.dist.tss = 2000,
  window.size = 500 * 10^3,
  num.flanking.genes = 5,
  rm.promoter.regions.from.distal.linking = TRUE
)

Arguments

regions.gr
A Genomic Ranges object (GRanges) or a SummarizedExperiment object (rowRanges will be used)

genome
Human genome of reference "hg38" or "hg19"

method
How genes are mapped to regions: region overlapping gene promoter ("genes.promoter.overlap"); or genes within a window around the region ("window"); or a fixed number genes upstream and downstream of the region ("nearby.genes"); or closest gene to the region ("closest.gene")

promoter.upstream.dist.tss
Number of base pairs (bp) upstream of TSS to consider as promoter regions. Defaults to 2000 bp.

promoter.downstream.dist.tss
Number of base pairs (bp) downstream of TSS to consider as promoter regions. Defaults to 2000 bp.

window.size
When method = "window", number of base pairs to extend the region (+- window.size/2). Default is 500kbp (or +/- 250kbp, i.e. 250k bp from start or end of the region)

num.flanking.genes
When method = "nearby.genes", set the number of flanking genes upstream and downstream to search.Defaults to 5. For example, if num.flanking.genes = 5, it will return the 5 genes upstream and 5 genes downstream of the given region.

rm.promoter.regions.from.distal.linking
When performing distal linking with method = "windows" or method = "nearby.genes", if set to TRUE (default), probes in promoter regions will be removed from the input.

Value

A data frame with the following information: regionID, Target symbol, Target ensembl ID

Details

For the analysis of probes in promoter regions (promoter analysis), we recommend setting method = "genes.promoter.overlap".

For the analysis of probes in distal regions (distal analysis), we recommend setting either method = "window" or method = "nearby.genes".

Note that because method = "window" or method = "nearby.genes" are mainly used for analyzing distal probes, by default rm.promoter.regions.from.distal.linking = TRUE to remove probes in promoter regions.

Examples

library(GenomicRanges)

#> Loading required package: stats4

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:dbplyr’:
#> 
#>     ident, sql

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Create example region
regions.gr <- data.frame(
       chrom = c("chr22", "chr22", "chr22", "chr22", "chr22"),
       start = c("39377790", "50987294", "19746156", "42470063", "43817258"),
       end   = c("39377930", "50987527", "19746368", "42470223", "43817384"),
       stringsAsFactors = FALSE)  %>%
     makeGRangesFromDataFrame

 # map to closest gene
 region.genes.promoter.overlaps <- get_region_target_gene(
                      regions.gr = regions.gr,
                      genome = "hg19",
                      method = "genes.promoter.overlap")

#> Mapping regions to the closest gene

 # map to all gene within region +- 250kbp
 region.window.genes <- get_region_target_gene(
                      regions.gr = regions.gr,
                      genome = "hg19",
                      method = "window",
                      window.size = 500 * 10^3
 )

#> Removing regions overlapping promoter regions

#> o Get promoter regions for hg19

#> o Remove promoter regions

#> Mapping regions to genes within a window of size: 5e+05 bp

 # map regions to n upstream and n downstream genes
 region.nearby.genes <- get_region_target_gene(
                      regions.gr = regions.gr,
                      genome = "hg19",
                      method = "nearby.genes",
                      num.flanking.genes = 5
 )

#> Removing regions overlapping promoter regions

#> o Get promoter regions for hg19

#> o Remove promoter regions

#> Identifying genes close to the region

#> Identifying 5 genes downstream to the region

#> Identifying 5 genes upstream of the region

#> Identifying gene position for each region