Get residuals from regression model

Compute studentized residuals from fitting linear regression models to expression values in a data matrix

get_residuals(
  data.matrix,
  metadata.samples = NULL,
  metadata.genes = NULL,
  cores = 1
)

Arguments

data.matrix
A matrix with samples as columns and features (gene, probes) as rows. Note that expression values should typically be log2(x+1) transformed before fitting linear regression models.

metadata.samples
A data frame with samples as rows and columns the covariates. No NA values are allowed, otherwise residual of the corresponding sample will be NA.

metadata.genes
A data frame with genes (covariates) as rows and samples as columns. For each evaluated gene, each column (e.g. CNA) that corresponds to the same gene will be set as a single covariate variable. This can be used to correct copy number alterations for each gene.

cores
Number of CPU cores to be used. Defaults to 1.

Value

A residuals matrix with samples as columns and features (gene, probes) as rows

Details

When only metadata.samples are provided, this function computes residuals for expression values in a data matrix by fitting model

features ~ Sample_covariate1 + Sample_covariate2 ... + Sample_covariateN where N is the index of the columns in the metadata provided, features are (typically log transformed) expression values.

When the user additionally provide metadata.genes, that is, gene metadata (e.g. gene_covariate = copy number variations/alterations) residuals are computed by fitting the following model:

features ~ Sample_covariate1 + Sample_covariate2 ... + Sample_covariateN + gene_covariate

Examples

data("gene.exp.chr21.log2")

data("clinical")
metadata <- clinical[,c( "gender", "sample_type")]

cnv <- matrix(
   sample(x = c(-2,-1,0,1,2),
   size = ncol(gene.exp.chr21.log2) * nrow(gene.exp.chr21.log2),replace = TRUE),
   nrow = nrow(gene.exp.chr21.log2),
   ncol = ncol(gene.exp.chr21.log2)
)
rownames(cnv) <- rownames(gene.exp.chr21.log2)
colnames(cnv) <- colnames(gene.exp.chr21.log2)

gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21.log2[1:3,],
   metadata.samples = metadata,
   metadata.genes = cnv
)

#> Formula used: val ~ gender + sample_type + gene.covariate

#> |                                                    |  0%                      |=================                                   |33.33333% ~0 s remaining  |==================================                  |66.66667% ~0 s remaining  |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s 

gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21.log2[1:3,],
   metadata.samples = metadata,
   metadata.genes = cnv[1:2,]
)

#> Formula used: val ~ gender + sample_type + gene.covariate

#> |                                                    |  0%                      |=================                                   |33.33333% ~0 s remaining  |==================================                  |66.66667% ~0 s remaining  |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s 

gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21.log2[1:3,],
   metadata.samples = metadata
)

#> Formula used: val ~ gender + sample_type

#> |                                                    |  0%                      |=================                                   |33.33333% ~0 s remaining  |==================================                  |66.66667% ~0 s remaining  |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s