Plot stratified model results

Create several plots to show interaction data TF expression with target gene interaction using a linear model $$log2(RNA target) ~ log2(TF)$$ to samples with highest DNAm values (top 25 percent) and lowest DNAm values (bottom 25 percent), separately.

plot_stratified_model(
  triplet.results,
  dnam,
  exp,
  metadata,
  tf.activity.es = NULL
)

Arguments

triplet.results
Output from function stratified_model with Region ID, TF (column name: TF), and target gene (column name: target), p-values and estimates of interaction

dnam
DNA methylation matrix (columns: samples same order as met, rows: regions/probes)

exp
gene expression matrix (columns: samples same order as met, rows: genes)

metadata
A data frame with samples as rownames and one columns that will be used to color the samples

tf.activity.es
A matrix with normalized enrichment scores for each TF across all samples to be used in linear models instead of TF gene expression.

Value

A ggplot object, includes a table with results from fitting stratified model, and the following scatter plots: 1) TF vs DNAm, 2) Target vs DNAm, 3) Target vs TF, 4) Target vs TF for samples in Q1 and Q4 for DNA methylation, 5) Target vs DNAm for samples in Q1 and Q4 for the TF

Examples

library(dplyr)
dnam <- runif(20,min = 0,max = 1) %>%
  matrix(ncol = 1) %>%  t
rownames(dnam) <- c("chr3:203727581-203728580")
colnames(dnam) <- paste0("Samples",1:20)

exp.target <-  runif(20,min = 0,max = 10) %>%
  matrix(ncol = 1) %>%  t
rownames(exp.target) <- c("ENSG00000232886")
colnames(exp.target) <- paste0("Samples",1:20)

exp.tf <- runif(20,min = 0,max = 10) %>%
  matrix(ncol = 1) %>%  t
rownames(exp.tf) <- c("ENSG00000232888")
colnames(exp.tf) <- paste0("Samples",1:20)

exp <- rbind(exp.tf, exp.target)

triplet <- data.frame(
   "regionID" =  c("chr3:203727581-203728580"),
   "target" = "ENSG00000232886",
   "TF" = "ENSG00000232888"
)

results <- stratified_model(triplet = triplet,dnam = dnam, exp = exp)

#> Removing genes with RNA expression equal to 0 for all samples from triplets

#> Removing triplet with no DNA methylation information for more than 25% of the samples

#> |                                                    |  0%                      |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s 

plots <- plot_stratified_model(
    triplet.results = results,
    dnam = dnam,
    exp = exp
)

#> |                                                    |  0%                      

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> |====================================================|100% ~0 s remaining       |====================================================|100%                      Completed after 0 s 

if (FALSE) {
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21.log2")
triplet <- data.frame(
   "regionID" = rownames(dna.met.chr21)[1:5],
   "TF" = rownames(gene.exp.chr21.log2)[11:15],
   "target" = rownames(gene.exp.chr21.log2)[1:5]
)
results <- stratified_model(
   triplet = triplet,
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2
)
plots <- plot_stratified_model(
   triplet.results = results[1,],
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2
)
# Adding color to samples
metadata <- clinical[,"sample_type",drop = FALSE]
plots <- plot_stratified_model(
   triplet.results = results[1,],
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2,
   metadata = metadata
)
}